requirements.txt
aiohttp==3.6.2
paho-mqtt==1.5.0
